Authentication/authorization added to API endpoints. Added style to SPA pages and created new components for login/register.

+++ app_admin/src/app/trip-card/trip-card.component.ts
@@ -3,6 +3,7 @@ import { Router } from "@angular/router";
 
 import { Trip } from '../models/trip';
 import { TripDataService } from '../services/trip-data.service';
+import { AuthenticationService } from '../services/authentication.service';
 
 /* 
   This component holds the view for an individual trip and allows 
@@ -17,9 +18,17 @@ import { TripDataService } from '../services/trip-data.service';
 export class TripCardComponent implements OnInit {
   
   @Input('trip') trip: Trip;
-  constructor(private router: Router, private tripDataService: TripDataService) { }
+  constructor(
+    private router: Router, 
+    private tripDataService: TripDataService,
+    private authenticationService: AuthenticationService
+  ) { }
 
   ngOnInit(): void {
+  }
+
+  public isLoggedIn(): boolean {
+    return this.authenticationService.isLoggedIn();
   }
 
   private editTrip(trip: Trip): void {
@@ -30,20 +39,23 @@ export class TripCardComponent implements OnInit {
     this.router.navigate(['edit-trip']);
   }
 
-  private deleteTrip(): void {
-    console.log('Inside TripLIstingComponent#deleteTrip');
+  private deleteTrip(trip: Trip): void {
+    console.log('Inside TripListingComponent#deleteTrip');
     // Confirm the delete of trip
     const confirmDelete = confirm("Are you sure you want to delete this trip?");
+    const isLoggedIn = this.isLoggedIn();
 
     if (confirmDelete) {
-      // Call deleteTrip method from the service
-      this.tripDataService.deleteTrip(this.trip.code)
-        .then(() => {
-          console.log("Trip successfully deleted");
-        })
-        .catch(err => {
-          console.log("Error deleting trip: " + err);
-        });
+      if (isLoggedIn) {
+        // Call deleteTrip method from the service
+        this.tripDataService.deleteTrip(this.trip.code)
+          .then(() => {
+            console.log("Trip successfully deleted");
+          })
+          .catch(err => {
+            console.log("Error deleting trip: " + err);
+          });
+      } 
     }
   }
